third_party/WebKit/LayoutTests/css3/device-adapt/opera/constrain-016-expected.txt


FAIL CSS Test: @viewport constrained - height deduced from width and aspect ratio. assert_equals: expected 10000 but got 800
